feat: introduce concept of rulesets and variations
For performance reasons, bundle up all rules for a feature flag or
segment into a single entity to reduce the number of database queries
needed.

Also introduce variations as a first-class entity, allowing feature
flags to have multiple variations rather than just a simple on/off
state. This also more clearly separates the concept of a feature flag's
state from the rules that determine which state is served.